# docs/swagger.yaml
basePath: /
definitions:
  models.Person:
    properties:
      address:
        type: string
      name:
        description: gorm.Model
        type: string
    type: object
  views.GetAllPeopleSwagger:
    properties:
      error: {}
      message:
        example: GET_SUCCESS
        type: string
      payload:
        items:
          $ref: '#/definitions/models.Person'
        type: array
      status:
        example: 200
        type: integer
    type: object
host: localhost:4000
info:
  contact:
    email: teguh.afdilla138@gmail.com
    name: Teguh Ridho Afdilla
  description: Sample API Spec for Orders
  termsOfService: http://swagger.io/terms/
  title: Orders API
  version: v1.0
paths:
  /people:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.GetAllPeopleSwagger'
      summary: Get all people
      tags:
      - person
    post:
      consumes:
      - application/json
      parameters:
      - description: Add New People
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Person'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.GetAllPeopleSwagger'
      summary: Add new people
      tags:
      - person
    put:
      consumes:
      - application/json
      parameters:
      - description: Add New People
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Person'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.GetAllPeopleSwagger'
      summary: Edit existing people
      tags:
      - person
  /people/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: People ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.GetAllPeopleSwagger'
      summary: Delete existing people
      tags:
      - person
swagger: "2.0"